code cleanup, some 'normal' commands
